[AI] Fix 'process is not defined' by separating client and server code
Root cause: customFunctions.ts was imported by client-side code but
had server dependencies, causing db module (and transitive dependencies
like adm-zip) to be bundled into browser.

Solution: Split into two files to properly separate concerns:
1. customFunctions.ts - Plugin only (no server imports, safe for browser)
2. customFunctionsPreferences.ts - Server-only preferences loading (uses aqlQuery)

Changes:
- Created customFunctionsPreferences.ts with loadUserPreferencesForFormulas()
- Uses aqlQuery (Actual's custom SQL) instead of direct db imports
- Moved helper functions (getLocaleDefaults, getCurrencyFromLocale) to new file
- customFunctions.ts now only has setCachedUserPreferences() setter
- action.ts loads prefs and caches them at module init
- Client code can still import CustomFunctionsPlugin safely
