+++ packages/loot-core/src/server/rules/action.ts
@@ -13,8 +13,9 @@ import { amountToInteger } from '#shared/util';
 import {
   CustomFunctionsPlugin,
   customFunctionsTranslations,
-  loadUserPreferencesForFormulas,
+  setCachedUserPreferences,
 } from './customFunctions';
+import { loadUserPreferencesForFormulas } from './customFunctionsPreferences';
 import { assert } from './rule-utils';
 
 HyperFormula.registerLanguage('enUS', enUS);
@@ -25,7 +26,9 @@ HyperFormula.registerFunctionPlugin(
 
 // Load user preferences on module initialization for formatting functions
 // This is async but happens once when the module loads
-void loadUserPreferencesForFormulas();
+void loadUserPreferencesForFormulas().then(prefs => {
+  setCachedUserPreferences(prefs);
+});
 
 const ACTION_OPS = [
   'set',
